requirements.txt
sh
pytest
pytest-xdist
path.py
